feat: import everything without errors
Full conversion of the Uluru subset we currently recognize into a
(simplified) model.

Big addition: "schema unification". Given two schema types `A` and
`B`, compute their "typed union". That is to say, if represent the
same type return that type with all assertions widened (and fail
if they don't represent the same type).

Everything currently converts successfully, which is atypical. Not sure
where the skeletons are buried yet...