feat: convert all registry resources, unifying schemas (#17)
Full conversion of the Registry subset we currently recognize into a
(simplified) model.

Big addition: "schema unification". Given two schema types `A` and `B`,
compute their "typed union". That is to say, if represent the same type
return that type with all assertions widened (and fail if they don't
represent the same type).

Everything currently converts successfully, which is atypical. Not sure
where the skeletons are buried yet...

Fixes #

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>
Co-authored-by: github-actions <github-actions@github.com>