chore: merge queue improvements
The merge queue should not require on the pr validation workflow, but we want it enforced on the PR itself.

Make PRs run the build workflow on the target like the merge queue does. This way PRs will always run builds against the latest workflow version and do not need to be kept up to date.