feat(cfn-resources): generate property interfaces (#41)
## Code generation

The AST is build using classes and spec structs.
The spec structs are currently re-using existing `@jsii/spec` to speed
things up, but types are intended to eventually be fully separate.
A renderer then takes care of converting the AST into code.

### Why no callbacks?

The callback pattern is effectively syntactic sugar to manipulate the
spec in well-defined scopes.
While I see a lot of potential to improve the ease of use, I don't
believe it's essential right now.
We are also still experimenting with the exact shape of things, so I'd
don't want to introduce unnecessary abstractions.

### Generated files are not included

We have discussed this in person.
While there some (weak) benefits to commit generated files, there are
also valid reasons against it.
However, it's much easier to add these files later than removing them
after the fact.
So until a further decision is taken, we do not commit generated files.

### New packages

- `@cdklabs/typewriter` contains the generic parts to build the AST and
render it as TS code
- `@aws-cdk/cfn-resources` contains the cli code to generate L1
Constructs from our database and the generated files

### Monorepo changes

- Introduces a vscode workspace file as an alternative usage mode
- To support this, some configuration needed to moved from `.projenrcts`
to the `MonorepoRoot` project type and made dynamic
- Small formatting changes to top-level markdown files