chore: use prettier everywhere, with the same rules (#47)
[138685b](https://github.com/cdklabs/awscdk-service-spec/commit/138685b5afd363e97df20964ed53c90934bea1d5)
added prettier rules at the top-level of the monorepo.
This causes IDEs to interpret these rules inconsistently depending on a
user's usage.
For example opening the repo root as project will cause VSCode to
complain about violations.
But opening a subproject directly will fallback to use the configured
eslint settings.
This is made worse that behavior may change based on existence of
plugins or.

This PR accepts that the prettier rules that had been configured at the
top-level are what we want to do.
It than makes sure that eslint and prettier are configured accordingly
at subproject level.
This will allow for more IDE scenarios to be handled consistently.
But more importantly, it will enforce consistent rules as part of build
& self-mutation.

Finally the PR makes all existing code comply with the new rules.