feat: support type references (#53)
Adds support for references to types that are defined elsewhere.

CloudFormation types can be recursive.
This is resolved by checking if the type already exists in the scope and
re-using the existing type as needed.

Refactoring AST construction to a class is required because we need
access to the database to resolve type references.

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>