feat: generate initial version of resource classes (#63)
Changes:

* Generate resource classes
* Add docstrings
* Generate forwards and backwards prop type mappers (trying to introduce
the terms "producers" and "parsers")
* Add proxies for generating expressions `$E(obj).foo.bar(x);` -> `new
Invocation(new ObjectAccess(new ObjectAccess(obj, 'foo'), 'bar'), x)`
* Make a distinction between `Statement` and `Expression`.
* Add many more statements and expressions (all in the same file, might
not be the best place/best names anymore)
* Make a distinction between a `Type` and a `TypeDeclaration`
* Introduce the notion of `Symbol` (in a class called `ThingSymbol`,
name could/should be better), which is a name that can be
imported/exported/referenced from one scope to another. Symbols aren't
attached to imports, but trying to reference a symbol in a scope in
which it's not visible (either by importing or because it's in there
already) leads to an exception.
* THIS IS NOT COMPLETE YET - above describes the vision, but this is not
fully implemented, and some things like local variables are simply an
"identifier" expression, not a symbol.
* Turn renderer from a stateless class that returns complete strings
into a stateful class that accumulates a string one token at a time.
* Copy over existing naming conventions from `cfn2ts`.

Obvious TODOs from this ream of code:

* Tests
* Property validation
* Comments on statements and expressions
* Tagging should be handled specially