feat: add tagging (#81)
Add support for tagging. Comes with the following changes:

* There are a number of problems with the incoming schema files. Add
patching for them, and report the patches and warnings to a file so that
we can easily report them to upstream resources.
* Tags are represented varyingly in the CloudFormation as a list of `[ {
Key, Value } ]` or map of `{ [Key]: Value }`. Account for that via
`tagType`.
* Tags are passed into the resource props typed as `tags?: cdk.CfnTag[]`
(introduce a new kind of type called a `builtInType` for that), but
exposed on the resource class as `public readonly tags = new
TagManager();` (fun!).

Resolves #89

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>