fix: emit property types in class scope (#103)
In the current cfn2ts, we emit all property types in the scope of a
resource class, so that their names don't conflict with each other.

The way that works in TypeScript is by following a `class` declaration
with a `namespace` declaration:

```ts
class SomeClass { ... }

namespace SomeClass {
  interface NestedType { ... }
}

// Refer by using:
let someVar: SomeClass.NestedType;
```

To support that, we make the following changes:

- The concept of a Scope is now an interface `IScope`. We leave have
`ScopeImpl` available as a default implementation.
- Separate the concept of an `Import` and a `ScopeLink`. An Import
*leads to* a Scope Link, but other things may also lead to a Scope Link.
- `ClassType` now implements `IScope`, and when a new type is registered
into it, it holds on to the subtype and registers a Scope Link into the
parent scope.
- Because of Scope Linking, the lookup mechanism already works
automatically. The final addition is the TypeScript emitter that knows
it must emit the `namespace` declaration when it finds nested types
inside a class.

This PR also contains a refactor of the CloudFormation -> AST builder
machinery, moving it from the `AstBuilder` class into a specific
`TypeConverter` class that operates in the context of a single resource
(makes it so we don't have to pass the `resourceClass` scope around
through a bunch of interconnected functions, but they can all access
`this.resourceClass`). In hindsight this wasn't 100% necessary, but I
like the cleanup this gives so left it this way.

Fixes #

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>