feat(cfn-resources): property types are resolvable (#106)
Introduces the concept of resolvable types.

A CHANGE compared to the `cfn2ts` codegen is that we treat resource
types and nested property types the same in relation to `IResolvable`.
This is only relevant in the internal `_fromCloudFormation` code path.
Previously top-level resource types did not check for `IResolvable` and
would result in an undefined error cases. Now we explicitly check for
this and throw an error. The reward is that the parser functions for
resource types and property types are now identical from a codegen
perspective.


New generic features introduced along the way:
- `yarn generate --service=aws-iam` to only codegen a single service,
useful for debugging
- `stmt.throw_`
-  Methods can now be static
- `Type.ambient()` to easily get an ambient type

Unrelated small-ish fixes:
- Some type validators had the wrong method name configured
- When adding `unionValidator` the mapping function that is going over
the union types did not have access to `this`
- `ResourceClass._fromCloudFormation` should have been static, but
wasn't
- `StrContact` -> `StrConcat`
- Removed unused `./runtime-builtin
- Removed `Type.builtInType` and related code in favor of `Type.ambient`
because it was confusing me and was unused (I suspect these were left
over artifacts)

Resolves #94

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>