feat: augmentations (#120)
Adds "augmentations" (a poor man's modeled metrics) to the database and
code generator.

An augmentation file is generated, which extends otherwise hand-written
interfaces and classes using TypeScript type declaration merging and
JavaScript monkey patching. In development mode, files with supposedly
hand-written classes that the augmentations extend are also generated so
that we can compile everything in this repository (controlled with
`--augmentations-support` flag).

The following things were also changed in the making of this PR:

Typewriter:

- Add the concept of "monkey patching" to the code generator. A
`MonkeyPatchedType` is a new type of declaration that the renderer
recognizes. It behaves like a class from the point of view of the code
model (you can add methods to it), but the renderer will render it
differently.
- Rename `SymbolKind` to `DeclarationKind` (because that's what it is)
and move it to `type-declaration.ts`.
- Add selective imports, and the concept of making a single symbol from
one scope visible in another.
- Add splat expressions, make `object` also take splats. `expr.object`
now no longer is a `{ string -> string }`, it is a `{ string ->
Expression }` and you need to make `Expressions` out of literal strings
in the value by calling `expr.lit()`.
- `new TypeProxy(...)` now returns an `ExpressionProxy` so you can chain
calls: `new TypeProxy(...).method(arg)` without having to wrap yourself.

TSKB:

- Change the fields `from` and `to` on following links in the database
graph to `entity`; seems more logical.
- Add a `get` field that can follow a `Reference` (so you don't have to
dereference the field yourself).
- Also add an `.only()` method to the resulting array from a `lookup`.

Fixes #96.