feat(tskb): allow indexing entities by an arbitrary mapping (#124)
Oh, the types! I hope you like angle brackets.

This massively changes how the types are done in `tskb`, with the end
goal of allowing indexing on arbitrary expressions rather than just
entity fields.

Here is the meat:

```ts
new Database({
  thing: entityCollection<Thing>().index({
    // "old" index
    name: fieldIndex('name', stringCmp),

    // "new" index
    lowercaseName: calculatedIndex((x) => x.name.toLowerCase(), stringCmp),
  }),
});
```

To achieve this, we had to drastically decouple the types from the
database schema definition, and rely more on inference. The bigger
changes are therefore:

- The type of the entire database is no longer defined, but instead
inferred from its construction.
- Database construction is now a two-step process: first we define the
entities, then the relationships between them. This is necessary to
build the relationships in a type-checked way, where we need to make
sure the entity set you're indicating is of the right type (and since we
don't have an a priori type definition, we need the value that
represents the universe of entities type-inferred already).

Fixes #

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>