fix(service-spec-sources): report output is not correct (#191)
The patch report wasn't correct for nested properties.
I tried fixing it, but ended up doing a bigger refactor.
It's mostly simplifying the data structure of `PatchReport` by means of
getting rid of the tree of lenses.
The reasoning here is that a JsonPatch is always applied to the root
document.
We already make sure to only report the first applied patch.
Therefore it should be enough to format the patch based on the root
document.
I might be missing something though.

It ends up with this kind of fix:

![image](https://user-images.githubusercontent.com/379814/233203467-51a475d7-1f1b-4a1b-b654-8a8fc71026cd.png)

As part of this I've added some basic test cases.
Also fixes inconsistent trailing commas (all are now removed).
The `canonicalize` fix was required because with a different tsconfig it
would not pass tests.