feat(service-spec-sources): support functional patches (#193)
Declarative Json Patch Operations are great when we are dealing with
stable and addressable input.
Specifically removing items form an array might not be a stable
operation and is potentially better implemented in code.

This change introduces an additional, optional functional interface for
patches.
It works by observing changes made to the document under the lens and
converting them into JsonPatch Operations (a library is doing this for
us).
These operations are then transparently re-introduced to the normal
flow.
The cool thing is that the fp-style patches can be fully typed.


Includes transplant of:
- `500_WAFv2_RuleGroup_Rename_Properties_patch.json`
- `510_ElasticSearch_Domain_patches.json`