fix(cfn-resources): timestamp properties are not of type Date (#213)
In the current spec, some properties are the primitive `Timestamp` type.
And in the L1s we convert these to native JS `Date`.
Unfortunately this information is not consistently available in the new
schema.
We are now looking this type up from the old spec and amend it to the
new schema.

Fixes #211