chore: build current aws-cdk-lib and compare spec with proposed new spec (#223)
Fetch the latest released version of `aws-cdk-lib` and compare its jsii
assembly with the proposed new spec.
Report and fail for any incompatible changes.

Note: This does not fail on external API changes yet. Needs a new jsii
release for that and an update to the `jsii-diff` call.
Also does the diff in a separate job as we expect it to fail just now
but want to enforce the compile step.