feat(service-spec-build): support preserving type history in spec db (#250)
For some types we need to use the original types for backwards
compatibility.
E.g. many properties have been converted from a generic `Json` type to a
well-defined struct.
However we also need to emit the type declarations for these new, unused
structs - since that's what we are doing currently in `cfn2ts`.
Therefore we now keep a history of all types.
We then resolve and add all types to the module, but use the original
type for the property itself.