fix(cfn-resources): make a distinction between 'legacy' and 'modern' taggability
Because of the way the old `cfn2ts` used to recognize and mark taggable
resources, not all taggable resources in fact implement the `ITaggable`
protocol properly.

Because of the way decisions had been made in the past, it is now
impossible to be both backwards compatible AND implement the taggability
protocol.

```ts
// ITaggable says
interface ITaggable {
  readonly tags: TagManager;
}

// But some classes already have
class CouldBeTaggableResource {
  public tags?: Array<CouldBeTaggableResource.MyTagProperty>;
}

// And there's no way to reconcile those 2 anymore
```

We need to target the old taggability protocol for old resources, and
the new taggability protocol for the new resources.

Needs https://github.com/aws/aws-cdk/pull/25610