fix(cfn-resources): complex types nested in unions are not resolvable (#264)
Recursively make complex types like arrays and maps resolvable when they
are part of a union. This is like it was in the old code gen.
Then use a distinctUnion to remove and flatten union of unions into a
single flat union with only distinct elements.

Fixes the underwhelming amount of 4 jsii-diff issues.