feat(service-spec-build): merge object type unions in references (#265)
A type reference might be a union of object types.
This is a new feature in the Registry Schema compared to the CFN Spec.
It's mostly used to create XOR type properties within a type: Either set
`a` or set `b`.
In future these should be represented as an enum-like class.
However to maintain backwards-compatibility we now have to interpret
this schema definition and merge the objects of the union into a single
object type.

Also includes a fix to to de-duplicate type references if they are the
same per resource.

Fixes 75 jsii-diff issues.