fix(cfn-resources): make a distinction between 'legacy' and 'modern' taggability (#256)
Because of the way the old `cfn2ts` used to recognize and mark taggable
resources, not all taggable resources in fact implement the `ITaggable`
protocol properly.

Because of the way decisions had been made in the past, it is now
impossible to be both backwards compatible AND implement the taggability
protocol.

```ts
// ITaggable says
interface ITaggable {
  readonly tags: TagManager;
}

// But some classes already have
class CouldBeTaggableResource {
  public tags?: Array<CouldBeTaggableResource.MyTagProperty>;
}

// And there's no way to reconcile those 2 anymore
```

We need to target the old taggability protocol for old resources, and
the new taggability protocol for the new resources.

Needs https://github.com/aws/aws-cdk/pull/25610

## Also in this PR

Because the mapping between CloudFormation property, PropsType property
and class property becomes more complex, I've decided to refactor the
logic a bit:

- `ResourceDecider`: based on the model, decides what properties to
generate for each type for resources
- `ResourceClass`: takes the output of `ResourceDecider` and builds the
appropriate `typewriter` objects
- `TypeDefinitionDecider`: the same as `ResourceDecider`, but for type
definitions
- `TypeDefinitionStruct`: takes the output of `TypeDefinitionDecider`
and builds the appropriate `typewriter` objects
- `TypeConverter`: now has a callback to build types for TypeDefinitions
(the default implementation uses `TypeDefinitionStruct`)
- `CloudFormationMapping`: a renamed version of `PropMapping`, is now
also in charge for building the proptype CFN producer and parser (these
used to be on `TypeConverter`).

Also in this PR: the properties are now sorted alphabetically, which
they didn't use to be. So reviewing the snapshot changes is a bit of a
bother. Sorry.

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>