feat(typewriter): statements & expressions can have comments (#286)
Resolves #98 

Differentiate between comments on statements (which are in a separate
line) and comments on expressions (which are inline).
Logic for expressions is very basic right now.
We can't do more complex stuff yet like comments on property
assignments, since these are not model as a separate expression and
rather as a single expression of type Object Literal.
Changing this will require a larger refactor.

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>