feat: load legacy spec into the database as `previousTypes` (#278)
Load the legacy resource spec into the database, and use it to infer
"previous types" for properties.

When generating code, currently we generate only for the oldest
available type; this makes us generate equivalent code for the old spec.
This new functionality subsumes the "load old `tag` information from the
legacy spec" and "patch specific old types to `json`" faetures, so the
code for those is removed.

To ensure a minimal diff with `aws-cdk`, we have to also copy the entire
patchset feature and the sets of JSON patches that repository was
applying.

## Also in this PR

- Solve some issues around union handling (deduplicating duplicate union
entries, and if a type has an `anyOf` but doesn't differ in any of its
constituent `type` fields then don't normalize).
- Add a new program where we can put scripts to query the generated
database.
- Add a built-in `tag` type and normalize around that. Custom
resource-specific type that look a suspicious amount like the built-in
`Tag` type get normalized to that (for backwards compatibility).
- If `patternProperties` and `additionalProperties: true` are both
present, ignore the latter one, it's probably there by accident.
- Copy over the `patch-set` feature from the `aws-cdk` repo, plus all of
its patches, to come up with an equivalent spec.
-

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>
Co-authored-by: Momo Kornher <kornherm@amazon.co.uk>