chore: release non-private packages to GitHub Releases  (#324)
Add monorepo release functionality. Works like this:

- The main project creates a `MonorepoRelease` component. Every every
package will register itself with this component to create a
`WorkspaceRelease` component.
- The main component manages two things:
- A `release:all` task, which will call the required steps for release
in topological order on all registered subprojects.
- A GitHub Workflow that will call `release:all`, and then add
publishing steps for each registered subproject.
- The `WorkspaceRelease` component .
- Each package gets a `bump` and `unbump` task. These are responsible
for sorting out everything in regards to versions.
- Because we only ever call `bump` and `unbump` in topological order,
they can limit their concern strictly locally. Let's say package B
depends on A. When B is "bumped", we know that A already has the all the
correct version information set. We can therefor use information from A
to populate B.
- Depending if a package is `private` or `public`, the `bump` task will
determine if a version bump is needed and create the respective files
and the changelog.
 
The `release:all` task is responsible for calling all required steps in
topological order (each per step):
- clean-up any existing `dist` directories
- `bump` sorts out versions
- `build` create a releasable artifact for all packages
- `unbump` to restore the HEAD state of the repo


The GitHub Workflow needs some finagling because I want to reuse the
logic of the `Publisher` that already generates workflows, but it makes
some assumptions on identifiers that no longer hold when there is more
than one package: we need to do some work to replace identifiers in the
publishing steps.


For versioning, we prefix all release tags with the package name, so
that every package can (in principle) have its own version:
`@scope/package@v1.2.3`

We have configured `standard-version` to only consider the workspace's
path, this limits the release logs for each package and does not create
a new version if nothing has changed.

---------

Co-authored-by: Rico Huijbers <rix0rrr@gmail.com>