fix(sam-spec): replace `AlexaSkillEvent.Variables` with `AlexaSkillEvent.SkillId` (#459)
~~The spec we use claims that AlexaSkillEvent has a Variables property,
but it does not:
https://docs.aws.amazon.com/serverless-application-model/latest/developerguide/sam-property-function-alexaskill.html~~

This change: https://github.com/awslabs/goformation/pull/613/files made
the spec match the SAM docs. The property is marked as optional, but it
is the only possible property; make it required. Our codegen will now
only match the strongest possible type in type unions (which the inputs
to function events are) and `AlexaSkillEvent` will never be chosen,
because it currently has no required properties. This results in the
`AlexaSkillEvent` being ignored during synthesis, which creates
templates with missing properties.

Making `SkillId` required makes the `AlexaSkillEvent` usable in CDK
apps.

---------

Co-authored-by: Rico Hermans <rix0rrr@gmail.com>