refactor: import progressively
We used to do the following to import both old and new specs:

- Load resource information from the CFN Registry (new spec), enhancing
  with information from the old spec if available.
- Then load the old spec, and patch some field types back to an old type
  if necessary.

This method is not very composable, and doesn't allow easily layering
even more information on top.

Instead, we refactor it here to a multi-phase process, where each new
spec that is loaded refines and adds information on top of information
that's already in the database.

- First import the legacy spec, populating as much of the database as
  we can.
- Then do a registry spec import, again populating as much of the
  database as we can.

If we detect that some property type updates are not reconcilable,  we
push the current type into the `previousTypes` array, and add a new one.