feat: diff spec databases (#596)
Adds a diffing tool to the service spec repository. Its goal is to print
the differences between 2 database files, in terms of services and their
properties added, removed and updated. By default, the diff is printed
in a nice ASCII art tree, but it can also be printed in JSON format by
passing the `--json` flag.

This diff tool will be used in the future to make sure that when we
refactor how the spec database is loaded, we can ensure that the
database before the refactor and after the refactor is the same (because
the diff will be empty).

We've decided that while the `tskb` project could potentially support a
generic database-level diff, it's probably easiest for now to just diff
at the application level (i.e., the code doing the diffing knows about
services, resources, properties and attributes).

* Diff types have been added to the `service-spec-types` package.
* The diff tool itself lives in the `service-spec-build` package.

NOTE: the ASCII tree print is currently broken, but we are merging this
early because we need the diffing functionality itself to validate other
changes.

---------

Co-authored-by: Kaizen Conroy <zen.conroy@gmail.com>