chore: stop updating CloudFormation Specification (#621)
Here's the reasoning:

Future changes are going to happen, and their primary source of truth is
going to be in the CloudFormation Registry schema. Those changes will
*also* be downconverted into Specification changes.

Not all of the things that can be expressed in Registry can be expressed
in Specification. Notably: union types cannot be expressed. The
downconverter will have to pick one of the sides of the union and
express that, and disregard the other. It is doing so in a stateless
manner, without regard to changes. Therefore, any change from singleton
to union type has the chance of introducing breaking changes.

Now that we have a "layered import", where we import the old resource
spec first and import the registry schema on top, there is actually a
value to freezing the resource spec in place: it is a record of history,
of types that once were and need to be preserved into the future for
backwards compatibility within CDK. Continuing to update the spec
actually runs counter to the goal of preserving the type history.

All in all, we are better off freezing the registry spec and continuing
to update the registry schema.

Fixes #

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>