refactor(service-spec-importers): combine build and sources into a single new package (#622)
We are preparing to release a new public package to import files on top
of an existing (possibly empty) database.
We settled on releasing a single new package that does all the work via
a CLI and small API.

This PR prepares the new package by merging the existing code from
`service-spec-build` and `service-spec-sources` into a single package.
Not further refactoring has been, except to resolve some dependency
issues.