feat(service-spec-importers): expose importers via DatabaseBuilder (#631)
Preparing the `service-spec-importers` package for a public release,
pt2.
This change replaces existing exports of everything with explicit
exports for:
- `DatabaseBuilder` a new class that helps importing data into a
database
- `DbDiff` a class that helps diff'ing to databases
- `ProblemReport` a class to collect & report problems encountered
during import

While this change is already enough to import additional specs on top of
an existing database, we will need one further change to decouple the
package from the source files in this repo.