feat(service-spec-importers): registry schema importers accept patches (#640)
We want to move the service patches from the generic importers package
to the specific model package. Generic patches stay with the importers.
To support this, the two registry schema importers now accept an
optional patcher.

With that done, we can move the (now unused) service patches to
aws-service-spec and use them in the implementation over there.

---------

Signed-off-by: github-actions <github-actions@github.com>
Co-authored-by: github-actions <github-actions@github.com>