fix(importers): `required` can never be set to `false` (#667)
Before merging, the CloudFormation registry importer tried to not emit
superfluous fields into the db. So it would not emit `required: false`
since the default value of `required` is already `false`.

When we started merging, that lack of emitting meant that any existing
value of `required: true` would never be overwritten with `required:
undefined` (because `undefined` isn't a value!).

Change this to have the importer always emit `requied: true|false`, so
we can correctly update the requiredness of this property.

The merging code itself takes care of canonicalizing `required: false`
to removing the property from the data structure, after merging.