fix(tskb): entities not returned from index (#663)
For non-unique indexes not all entities were are returned during a
lookup. This was because our previous binary search implementation was
optimized for speed, and returning the index of an element within the
partition. This is enough when adding values into the SortedMap. However
when retrieving values, we need to find the lower bound of the
partition, so we can return all elements.

Changed the binary search implementation to the one from the C++
standard library.


Includes support for `OptionalStringIndexLookups` which is how I
discovered the bug.