fix(aws-service-spec): scrutinies missing for some policy document properties (#670)
We are checking if a property is a policy document in order to decide if
it is scrutinizable or not.
This check used to check if the property type was `json`.
However more complex types can also be used for policy documents.

This PR changes to check to test if `json` is assignable to the property
type.
In practice this means the property type can be a union that contains
the `json` type.