chore: don't diff db on main (#781)
Makes the currently "failing" build on main pass by bypassing the
`diff-db` job that is not useful on `main`.
With this we can actually see if everything is alright on main by
looking at the status of the latest commit.

<img width="615" alt="image"
src="https://github.com/cdklabs/awscdk-service-spec/assets/379814/9c7d6683-e747-4809-9244-a5f300faf223">