chore: fix union errors while importing cloudformation registry resources (#1064)
This [PR](https://github.com/cdklabs/awscdk-service-spec/pull/278)
changed the logic in `canonicalizeTypeOperators` to no longer lift out
`anyOf` or `oneOf` type operators that are embedded in type objects. As
a result, the current logic used in `jsonschema.anyOf` and
`jsonschema.oneOf` was too loosely defined and would return `true` for
type objects with an embedded `anyOf` or `oneOf` type operator. For
example:
```
"StorageConfiguration" : {
  "type" : "object",
  "description" : "The vector store service in which the knowledge base is stored.",
  "properties" : {
    "Type" : {
      "$ref" : "#/definitions/KnowledgeBaseStorageType"
    },
    "OpensearchServerlessConfiguration" : {
      "$ref" : "#/definitions/OpenSearchServerlessConfiguration"
    },
    "PineconeConfiguration" : {
      "$ref" : "#/definitions/PineconeConfiguration"
    },
    "RdsConfiguration" : {
      "$ref" : "#/definitions/RdsConfiguration"
    }
  },
  "required" : [ "Type" ],
  "oneOf" : [ {
    "required" : [ "OpensearchServerlessConfiguration" ]
  }, {
    "required" : [ "PineconeConfiguration" ]
  }, {
    "required" : [ "RdsConfiguration" ]
  } ],
  "additionalProperties" : false
}
```
As a result, the logic in `schemaTypeToModelType` treats the above type
object as a `oneOf` type operator instead of an object type and begins
recursing on the sub-schemas defined in the`oneOf` array. However, this
recursion fails since `{ required: [ property ] }` is not a defined type
and should have been handled here prior to recursing on the type object
properties:
https://github.com/cdklabs/awscdk-service-spec/blob/f266e9b36ece6f423bfd91548696b692982483f8/packages/%40aws-cdk/service-spec-importers/src/importers/import-cloudformation-registry.ts#L71

This PR introduces a fix to this by making `jsonSchema.oneOf`,
`jsonschema.anyOf`, and `jsonschema.AllOf` less permissive by also
checking that a possible `oneOf`, `anyOf`, or `allOf` type operator does
not also have a defined type. Note that while `allOf` wasn't directly a
problem since it is embedded in `oneOf` or `anyOf` type operators, we
should still enforce checking that a possible `allOf` doesn't have a
defined type.

---------

Signed-off-by: Francis <colifran@amazon.com>