chore: patch for AWS::SecurityLake::Subscriber `Source` (#1231)
The `Sources` property that is part of `AWS::SecurityLake::Subscriber`
resource is being dropped due to errors importing the `Source` type.
Since `Sources` is an array of `Source` elements, `Sources` is dropped
from the final db. This PR fixes the schema format of `Source` to
prevent errors while importing. The database diff below shows that the
`Sources` property will be added. This
[CR](https://github.com/cdklabs/awscdk-service-spec/pull/1118) shows
when the `Sources` property was removed but can be a reference to what
we should expect to see in the database diff for `Sources`.

---------

Signed-off-by: Francis <colifran@amazon.com>