chore: fail downloading sources on http status code != `200` (#1331)
Our [feat(sources): update
stateful-resources](https://github.com/cdklabs/awscdk-service-spec/pull/1106)
PR is currently failing to build:

```console
👾 build » compile » build:db | ts-node build/build-db.ts
FAIL


SyntaxError: Unexpected non-whitespace character after JSON at position 3
```

This happens because the `StatefulResources.json` file got updated with
an invalid value. Our workflow updates this with the contents of
`StatefulResources.json` coming from
[cfn-lint](https://github.com/aws-cloudformation/cfn-lint). However, In
https://github.com/aws-cloudformation/cfn-lint/pull/3257, this file was
[removed](https://raw.githubusercontent.com/aws-cloudformation/cfn-lint/main/src/cfnlint/data/AdditionalSpecs/StatefulResources.json)
(it is now inlined inside the python file that was using it). This
results in the file being updated with the string `404: Not Found` -
which is not a valid JSON.

In this scenario, we shouldn't have even created the PR, but rather fail
during the workflow execution. We didn't because our workflow ignores
http status codes when downloading a source, this PR fixes that.

### Notes

- This replaces build failure with workflow execution failure, but we
still need to figure out how to handle the file removal.